requirements.txt
PyGithub
discord.py
python-dotenv
beautifulsoup4
requests 